+++ app/models/squad_member.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Integer, String
+from sqlalchemy import BigInteger, Integer, String
 from sqlalchemy.orm import Mapped, mapped_column
 from app.database import Base
 
@@ -10,4 +10,4 @@ class SquadMember(Base):
     user_id: Mapped[int] = mapped_column(Integer, nullable=False, index=True)
     rank: Mapped[str] = mapped_column(String(8), nullable=False)  # E, D, C, B, A, S
     stars: Mapped[int] = mapped_column(Integer, default=0)  # 0-5
-    base_power: Mapped[int] = mapped_column(Integer, default=0)
+    base_power: Mapped[int] = mapped_column(BigInteger, default=0)
